+++ Weapons/ShotgunGenerator.cs
